Load browser config. data during initialization
Previously, data describing the available browsers was not loaded until
the web server handled a request which required it. Subsequent requests
were intended to re-use the result of this operation, persisted as
global application state.

That strategy was susceptible to a race condition: if the initial
request that triggered data loading was quickly followed by a similar
request, the loading logic would be executed a second time, and the
resuting in-memory representation would include two entries for each
available browser.

Re-factor the internals to load the browser data at module
initialization time (e.g. before any requests are received).