run.py: Re-implement results validation
Previously, `run.py` was extended to ensure that the actual result set
described the expected number of tests (within some threshold). This was
implemented via the `--list-tests` flag provided by the WPT CLI.

Unfortunately, this functionality requires that the web browser under
test be installed on the local system. This is not the case when running
tests remotely using WebDriver, so the new validation logic would fail
in those cases.

The same information can be retrieved from a previously-unused output
stream (the "raw log") provided by WPT during test running.

Re-implement results validation to instead inspect the structured output
of the test run itself.