Defer results validation until upload
This project enforces a 98% "completeness" threshold for results sets.
(Any dataset containing results for fewer than 98% of the tests expected
to be run in a given revision of WPT are not published online.)

Previously, when test results were collected as part of a single
process, this condition was enforced one for the full data set.

When the system was distributed across many machines, the condition was
enforced for each subset of tests. This change made the criteria for
uploading more strict because failure in a single subset could block
uploading of the entire results set, even if the overall data set still
contained greater than 98% of the expected results.

Re-implement the criteria to once again be enforced a single time, in
terms of the final result set.