Report on any incomplete result sets
Previously, datasets which omitted results for some tests were
considered acceptable and subsequently uploaded to the web. The
underlying issues which motivated that design have been resolved, so the
project may be in a position to reject any incomplete datasets.

However, due to the imprecise nature of the previous validation
strategy, the actual completeness levels for certain browser
configurations are not well understood. Specifically, reports for
browsers which are mediated by the Sauce Labs service may continue to be
incomplete.

Decrease the threshold for missing tests to zero percent during results
validation. Maintain the prior threshold of two percent during results
uploading.

This configuration will surface any persisting issues with
incompleteness in the build system's web interface without interfering
with the overall reporting behavior. In this way, project maintainers
can verify that missing results are now a truly anomalous case, and a
subsequent commit can reject such datasets outright.