Use multiple Sauce Labs user accounts
The Sauce Labs service enforces rate limits on requests to its HTTP API.
In order to avoid exceeding these limits, this project has previously
limited the number of concurrent builds through that service to four.

The rate limits are enforced on a per-user-account basis. Assign a
unique Sauce Labs user account to each of 20 build workers and allow
these workers to operate in parallel.