Store intermediate results on separate partition
When results uploading fails, the results are persisted on disk in order
to support manual recovery. Previously, the data was saved to the root
partition. This made it possible for multiple failed upload attempts to
exhaust system resources and interfere with related but distinct
services such as the web interface.

Reconfigure the build master to store intermediate results on a
dedicated partition (previously reserved for the Buildbot database).