Decrease logging verbosity
When running reftests, the "TBPL" logger prints the Base64 encoded
screen captures to the console. Although logs are eventually stored on
the build master, they are initially created within each worker. When
running large numbers of reftests, these logs occasionally exceed the
capacity of the build works.

Use the default "mach" logger which omits this data.