Select latest WebDriver binaries at execution time
Previously, this project installed statically-defined WebDriver server
binaries as part of the system provisioning process. New releases of
those binaries would only be used following a modification to the
project's provisioning scripts and a subsequent redeployment. While this
made the state of the system more clear, it also allowed for the
binaries to fall out of date.

Together with the web browsers, the WebDriver implementations are part
of the software under test. Due to this, it is prudent to use similar
procedures to identify and install both components.

Relocate the logic for downloading and installing WebDriver server
binaries from the system provisioning scripts to the test execution
process. Where possible, reuse code originally implemented to download
and install web browsers.