Accommodate secretive update from Sauce Labs
On 2018-08-09, the Sauce Connect binary published as version 4.5.0 was
modified, and as a result, it no longer matched the documented checksum
[1]. On 2018-08-10, Sauce Labs responded to an issue report by updating
the checksum but maintaining the new binary as version 4.5.0 [2].

This project takes a conservative approach to dependency management,
updating third-party utilities only when doing so is expected to improve
the system in some way. In this case, because the original version of
the binary is no longer available, the new version must be accepted.

[1] https://webcache.googleusercontent.com/search?q=cache:gj6omCtn5w8J:https://wiki.saucelabs.com/display/DOCS/2018/08/07/Announcing%2BSauce%2BConnect%2BProxy%2B4.5.0+&cd=1&hl=en&ct=clnk&gl=us
[2] https://wiki.saucelabs.com/display/DOCS/2018/08/07/Announcing+Sauce+Connect+Proxy+4.5.0