Add a flag to make timestamps optional
If --timestamps-optional is given, time_start & time_end become
optional. They will be omitted from the consolidated report if
time_start *or* time_end is missing in all chunks.

In addition to adding a new test for this flag, some old tests are
fixed: they used different port number for the server and the client.
test_duplicated_results is removed as this is no longer checked during
upload (the failure wasn't noticed when the behaviour was changed
because of the different port numbers).