Delay builds to 5 minutes past the hour
The wpt.fyi revision announcer selects new revisions at the hour
boundaries of the UTC clock. The results-collection project currently
queries at this same moment, making it prone to race conditions.

Update the scheduling logic to delay builds to 5 minutes past the hour
in order to minimize the influence of network delays and clock drift.