Re-enable SSL bumping during Sauce Labs testing
SSL bumping interferes with WebSocket traffic. This is why the feature
was originally disabled.

Running WPT locally depends on the use of self-signed certificates. SSL
bumping avoids the errors that self-signed certificates would otherwise
trigger in the browser under test.

Due to this situation, it is not currently possible to correctly test
both WebSocket behavior and HTTPS behavior. The latter should be
preferred because it is fundamental to the web platform.