Remove "wrapper" script
Under some circumstances, the WPT CLI kills the Sauce Connect process
with a "kill" operating system signal. When the process is created via
an intermediary "wrapper" script (such as the one implemented by this
project), the signal does not reach its intended target.

The WPT CLI recently implemented a feature to allow users to specify
options for the Sauce Connect binary [1]. Utilize that feature and
remove the intermediary script.

[1] https://github.com/web-platform-tests/wpt/pull/14356