Cancel collection from Firefox and Chrome
The web-platform-tests project has delegated the task of results
collection from Firefox and Chrome to the Taskcluster service. It is no
longer necessary for this project to collect results from those
browsers.

Decommission the infrastructure which was dedicated solely to collection
from Firefox and Chrome and persist machines which are used to collect
from Edge and Safari via Sauce Labs. Ensure that the Buildbot master
does not schedule further collection effort from Firefox and Chrome.