Correct virtual display settings
This project creates an X virtual frame buffer using the `xvfb-run`
wrapper script. That script uses default values for the screen's width,
height, and depth (640, 480, and 16, respectively). Those values violate
the basic assumptions of WPT [1] which include the following assertion:

> - The device has a viewport width of at least 800px.

In WPT's integration with the Taskcluster service, an X virtual frame
buffer is started directly via the `Xvfb` executable, and the screen
parameters are specified through environment variables [2].
`SCREEN_WIDTH`, `SCREEN_HEIGHT`, and `SCREEN_DEPTH` are set to 1280,
1024, and 24, respectively [3].

Configure this project to explicitly set the parameters of the virtual
display to use acceptable values. Select the same values in use by the
Taskcluster environment in order to promote parity between the two
systems.

[1] https://web-platform-tests.org/writing-tests/assumptions.html
[1] https://github.com/web-platform-tests/wpt/blob/83d93704f0cb36dd9fa3f45dd3e4b34990e934e6/tools/ci/run_tc.py#L142-L147
[2] https://github.com/web-platform-tests/wpt/blob/83d93704f0cb36dd9fa3f45dd3e4b34990e934e6/tools/docker/Dockerfile#L62-L64